refactor: consolidate constants and improve test imports
