+++ src/hooks/useMovieFetch.list.test.ts
@@ -3,7 +3,7 @@ import { http, HttpResponse } from "msw";
 import { describe, expect, it, vi } from "vitest";
 import { server } from "../test/server";
 import { mockMovie, mockMoviesPage1, mockMoviesPage2 } from "../test/handlers";
-import { TTL_MS } from "../test/constants";
+import { TTL_MS } from "../helpers";
 import type { MoviePropTypes } from "../Global.props";
 import type { MoviesState } from "./props";
 import useMovieFetch from "./useMovieFetch";
