--- src/test/constants.ts
@@ -1,1 +1,0 @@
-export const TTL_MS = 24 * 60 * 60 * 1000;
